projects/agent/pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.mypy]
files = ["pi_blink", "tests"]
pretty = true
python_version = "3.11"
show_error_codes = true
disable_error_code = "import-untyped"

[tool.poetry]
authors = ["parnell <152523161+leeparnell@users.noreply.github.com>"]
description = "Qai Agents for the Qai AI Platform"
name = "qai-agent"
packages = [{from = "src", include = "qai"}]
readme = "README.md"
version = "0.5.2.5"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
install = "^1.3.5"
llama-index-agent-openai = "^0.2.5"
llama-index-core = "^0.10.37"
llama-index-embeddings-huggingface = "^0.2.0"
llama-index-embeddings-openai = "^0.1.7"
llama-index-llms-openai = "^0.1.19"
llama-index-llms-replicate = "^0.1.3"
llama-index-program-openai = ">=0.1.5"
llama-index-tools-database = "^0.1.3"
llama-index-tools-google = "^0.1.3"
llama-index-vector-stores-mongodb = "^0.1.4"
multiprocess = "^0.70.16"
mysql-connector-python = "^8.4.0"
openai = "^1.23.2"
pandas = "^2.2.1"
pi-conf = "^0.7.7.5"
pi-log = "^0.5.8.1"
pip = "^24.0"
pydantic = "^2.7.0"
pymongo = "^4.6.3"
python-dateutil = "2.8.2"
python-dotenv = "^1.0.1"
qai-ai = {develop = true, path = "../ai"}
qai-core = {develop = true, path = "../core"}
pdl-api = { path = "../../../../pdl_api/pdl-api", develop = true }
sqlalchemy = "^2.0.29"
pi-blink = "^1.0.3"
instructor = "^1.3.2"

[tool.poetry.group.dev.dependencies]
accelerate = "^0.29.3"
bitsandbytes = "^0.42.0"
toml-sort = "^0.23.1"
transformers = "^4.40.1"
pytest-asyncio = "^0.23.7"
types-requests = "^2.32.0.20240602"

[[tool.poetry.source]]
name = "pytorch-gpu-src"
priority = "explicit"
url = "https://download.pytorch.org/whl/cu118"

[[tool.poetry.source]]
name = "pytorch-cpu-src"
priority = "explicit"
url = "https://download.pytorch.org/whl/cpu"

[tool.tomlsort]
all = true
in_place = true
overrides."tool.poetry.dependencies".first = ["python"]
